requirements.txt
# CLI dependencies (existing)
requests>=2.31.0
sqlite-utils>=3.36
rich>=13.7.0
jinja2>=3.1.3
weasyprint>=62.0
anthropic>=0.40.0
python-dotenv>=1.0.0
schedule>=1.2.0
click>=8.1.7
pandas>=2.2.0

# API dependencies (new)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pyjwt>=2.8.0
bcrypt>=4.1.0
pydantic-settings>=2.1.0
apscheduler>=3.10.0
stripe>=7.0.0
pymupdf>=1.23.0
resend>=0.7.0
httpx>=0.27.0
